Ports the compiler from Python 2.7 to Python 3.5

+++ Example/Bridge/Java/allogen/example/AsyncTaskCallback2.java
@@ -9,5 +9,5 @@
 package allogen.example;
 
 public interface AsyncTaskCallback2 {
-	public int onCallback(int a, int b);
+	public int onCallback(short a, short b);
 }
